Create highlight helper file
