+++ packages/autocomplete-client/src/search/fetchMeilisearchResults.ts
@@ -8,7 +8,9 @@ import {
   HITS_PER_PAGE,
 } from '../constants/index.js'
 import type { SearchClient as MeilisearchSearchClient } from '../types/SearchClient.js'
-import type { HighlightResult } from 'algoliasearch-helper/types/algoliasearch.js'
+import { FieldHighlight } from 'instantsearch.js/es/types/algoliasearch'
+import { calculateHighlightMetadata } from './highlight.js'
+import { mapOneOrMany } from '../utils.js'
 
 interface SearchParams {
   /** The initialized Meilisearch search client. */
@@ -22,13 +24,6 @@ interface SearchParams {
       }
     }
   >
-}
-
-interface HighlightMetadata {
-  value: string
-  fullyHighlighted: boolean
-  matchLevel: 'none' | 'partial' | 'full'
-  matchedWords: string[]
 }
 
 export function fetchMeilisearchResults<TRecord = Record<string, any>>({
@@ -86,7 +81,7 @@ export function fetchMeilisearchResults<TRecord = Record<string, any>>({
                     )
 
                     return acc
-                  }, {} as HighlightResult<TRecord>),
+                  }, {} as FieldHighlight<TRecord>),
                 }
 
                 // Attach metadata to each hit if present (for Meilisearch Cloud Analytics)
@@ -105,63 +100,7 @@ export function fetchMeilisearchResults<TRecord = Record<string, any>>({
     )
 }
 
-/**
- * Calculate the highlight metadata for a given highlight value.
- *
- * @param query - The query string.
- * @param preTag - The pre tag.
- * @param postTag - The post tag.
- * @param highlightValue - The highlight value response from Meilisearch.
- * @returns The highlight metadata.
- */
-function calculateHighlightMetadata(
-  query: string,
-  preTag: string,
-  postTag: string,
-  highlightValue: string
-): HighlightMetadata {
-  // Extract all highlighted segments
-  const highlightRegex = new RegExp(`${preTag}(.*?)${postTag}`, 'g')
-  const matches: string[] = []
-  let match
-  while ((match = highlightRegex.exec(highlightValue)) !== null) {
-    matches.push(match[1])
-  }
-
-  // Remove highlight tags to get the highlighted text without the tags
-  const cleanValue = highlightValue.replace(
-    new RegExp(`${preTag}|${postTag}`, 'g'),
-    ''
-  )
-
-  // Determine if the entire attribute is highlighted
-  // fullyHighlighted = true if cleanValue and the concatenation of all matched segments are identical
-  const highlightedText = matches.join('')
-  const fullyHighlighted = cleanValue === highlightedText
-
-  // Determine match level:
-  // - 'none' if no matches
-  // - 'partial' if some matches but not fully highlighted
-  // - 'full' if the highlighted text is the entire field value content
-  let matchLevel: 'none' | 'partial' | 'full' = 'none'
-  if (matches.length > 0) {
-    matchLevel = cleanValue.includes(query) ? 'full' : 'partial'
-  }
-
-  return {
-    value: highlightValue,
-    fullyHighlighted,
-    matchLevel,
-    matchedWords: matches,
-  }
-}
-
-// Helper to apply a function to a single value or an array of values
-function mapOneOrMany<T, U>(value: T | T[], mapFn: (value: T) => U): U | U[] {
-  return Array.isArray(value) ? value.map(mapFn) : mapFn(value)
-}
-
-type DefinedHighlightResult = { value: string } | Array<{ value: string }> // if the field is an array
+type DefinedHighlightResult = { value: string } | Array<{ value: string }>
 
 /**
  * Some fields may not return a value at all - nested arrays/objects for example
